캘린더뷰 추가하여 메인뷰 정리
