+++ umm/umm/Calendar.swift
@@ -1,0 +1,123 @@
+//
+//  Calendar.swift
+//  umm
+//
+//  Created by 강진 on 6/5/25.
+//
+
+import SwiftUI
+
+struct CalendarView: View {
+    let calendar: Calendar
+    @Binding var selectedDate: Date
+    @Binding var showDatePicker: Bool
+    @Binding var weekOffset: Int
+  
+    init(calendar: Calendar, selectedDate: Binding<Date>, showDatePicker: Binding<Bool>, weekOffset: Binding<Int>) {
+        self.calendar = calendar
+        self._selectedDate = selectedDate
+        self._showDatePicker = showDatePicker
+        self._weekOffset = weekOffset
+        UIDatePicker.appearance().tintColor = UIColor.orange
+    }
+
+    private var week: [Date] {
+        let startOfWeek = calendar.date(byAdding: .weekOfYear, value: weekOffset, to: calendar.startOfWeek(for: selectedDate))!
+        return (0..<7).compactMap { calendar.date(byAdding: .day, value: $0, to: startOfWeek) }
+    }
+
+    var body: some View {
+        VStack(spacing: 8) {
+            // 상단: 현재 월 표시 및 달력 버튼
+            HStack {
+                Spacer()
+                Text(DateFormatter.customMonthFormatter.string(from: week.first ?? selectedDate))
+                    .font(.system(size: 17))
+                Spacer()
+                Button(action: {
+                    showDatePicker = true
+                }) {
+                    Image(systemName: "calendar")
+                        .foregroundColor(.orange)
+                }
+            }
+            .padding(.horizontal)
+            .padding(.top)
+
+            // 요일 및 날짜 표시 영역
+            HStack {
+                ForEach(week, id: \.self) { date in
+                    let isSelected = calendar.isDate(date, inSameDayAs: selectedDate)
+                    VStack {
+                        Text(date, format: .dateTime.weekday())
+                            .foregroundColor(isSelected ? .orange : .gray)
+                            .font(.system(size: 12))
+                        Text(date, format: .dateTime.day())
+                            .foregroundColor(isSelected ? .orange : .primary)
+                            .font(.system(size: 17))
+                    }
+                    .frame(maxWidth: .infinity)
+                    .onTapGesture {
+                        selectedDate = date
+                        weekOffset = 0
+                    }
+                }
+            }
+            .padding()
+            .gesture(
+                DragGesture().onEnded { value in
+                    if value.translation.width < -20 {
+                        weekOffset += 1
+                    } else if value.translation.width > 20 {
+                        weekOffset -= 1
+                    }
+                }
+            )
+
+            Spacer()
+        }
+        .overlay(
+            Group {
+                if showDatePicker {
+                    ZStack {
+                        Color.white.opacity(0.01)
+                            .ignoresSafeArea()
+                            .onTapGesture {
+                                showDatePicker = false
+                            }
+
+                        VStack {
+                            DatePicker(
+                                "",
+                                selection: $selectedDate,
+                                displayedComponents: [.date]
+                            )
+                            .datePickerStyle(.graphical)
+                            .padding()
+                            .background(
+                                RoundedRectangle(cornerRadius: 16)
+                                    .fill(Color.orange.opacity(0.1))
+                            )
+                            .padding()
+                        }
+                    }
+                    .transition(.opacity)
+                }
+            }
+        )
+    }
+}
+
+extension Calendar {
+    func startOfWeek(for date: Date) -> Date {
+        self.dateInterval(of: .weekOfYear, for: date)!.start
+    }
+}
+
+extension DateFormatter {
+    static let customMonthFormatter: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "yyyy.MM"
+        return formatter
+    }()
+}
